Arreglar posicion fichas Reorganizar codigo Añadir restricciones de movimiento al rey

+++ app/Http/Controllers/Api/Ajedrez/Controllers/TableroController.php
@@ -3,16 +3,17 @@
 namespace App\Http\Controllers\Api\Ajedrez\Controllers;
 
 use Illuminate\Http\Request;
-use App\Http\Controllers\Api\Ajedrez\Master;
+use App\Http\Controllers\Controller;
+use App\Http\Controllers\Api\Ajedrez\Movimientos;
 use App\User;
 use App\Partida;
 use App\Ficha;
 
-class TableroController extends Master
+class TableroController extends Controller
 {
     function ver(Request $request){
-        $userID1 = $this->getIdUserFromToken($request->input('token'));
-        $userID2 = $this->getIdUserFromName($request->input('name'));
+        $userID1 = User::getIdUserFromToken($request->input('token'));
+        $userID2 = User::getIdUserFromName($request->input('name'));
 
         $status = 0;
         if($userID1 != false && $userID2 != false){
@@ -32,8 +33,8 @@ class TableroController extends Master
     }
 
     function moverFicha(Request $request){
-        $userID1 = $this->getIdUserFromToken($request->input('token'));
-        $userID2 = $this->getIdUserFromName($request->input('name'));
+        $userID1 = User::getIdUserFromToken($request->input('token'));
+        $userID2 = User::getIdUserFromName($request->input('name'));
         $toFila = $request->input('toFila');
         $toColumna = $request->input('toColumna');
         $fromFila = $request->input('fromFila');
@@ -53,35 +54,37 @@ class TableroController extends Master
                    ($partida->turno === "b" && $partida->id_jugador_blanco == $userID1)){
                     $ficha = Ficha::where([["id_partida", $partida->id], ["fila", $toFila], ["columna", $toColumna], ["color", $partida->turno]]);
                     if($ficha->count() == 1){
-                        $fichaTarget = Ficha::where([["id_partida", $partida->id], ["fila", $fromFila], ["columna", $fromColumna]]);
+                        if(Movimientos::checkMovimiento($ficha->first(), ["columna" => $fromColumna, "fila" => $fromFila])){
+                            $fichaTarget = Ficha::where([["id_partida", $partida->id], ["fila", $fromFila], ["columna", $fromColumna]]);
 
-                        if($fichaTarget->count() > 0){
-                            $fichaMia = $fichaTarget->first()->color === $partida->turno;
-                            if(!$fichaMia) $fichaTarget->delete();
-                        }else $fichaMia = false;
+                            if($fichaTarget->count() > 0){
+                                $fichaMia = $fichaTarget->first()->color === $partida->turno;
+                                if(!$fichaMia) $fichaTarget->delete();
+                            }else $fichaMia = false;
 
-                        //Fin partida
-                        if(Ficha::where([["id_partida", $partida->id], ["color", ($partida->turno === "n" ? "b" : "n")], ["tipo", "rey"]])->count()==0){
-                            $status = 2;
-                            $mensaje = "Fin partida, el ganador es el jugador ". ($partida->turno === "b" ? "blanco" : "negro");
+                            //Fin partida
+                            if(Ficha::where([["id_partida", $partida->id], ["color", ($partida->turno === "n" ? "b" : "n")], ["tipo", "rey"]])->count()==0){
+                                $status = 2;
+                                $mensaje = "Fin partida, el ganador es el jugador ". ($partida->turno === "b" ? "blanco" : "negro");
 
-                            Ficha::where("id_partida", $partida->id)->delete();
-                            $partida->delete();
+                                Ficha::where("id_partida", $partida->id)->delete();
+                                $partida->delete();
 
-                        //Mover Ficha
-                        }else if(!$fichaMia){
-                            $status = 1;
+                            //Mover Ficha
+                            }else if(!$fichaMia){
+                                $status = 1;
 
-                            $ficha = $ficha->first();
-                            $ficha->columna = $fromColumna;
-                            $ficha->fila = $fromFila;
-                            $ficha->save();
+                                $ficha = $ficha->first();
+                                $ficha->columna = $fromColumna;
+                                $ficha->fila = $fromFila;
+                                $ficha->save();
 
-                            $partida->turno = ($partida->turno === "n" ? "b" : "n");
-                            $partida->save();
+                                $partida->turno = ($partida->turno === "n" ? "b" : "n");
+                                $partida->save();
 
-                            $mensaje="ficha movida";
-                        }else $mensaje="No puedes mover tu ficha a un lugar donde ya hay otra ficha tuya";
+                                $mensaje="ficha movida";
+                            }else $mensaje="No puedes mover tu ficha a un lugar donde ya hay otra ficha tuya";
+                        }else $mensaje="Movimiento no valido.";
                     }else if($ficha->count() > 1) $mensaje="Se ha encontrado mas de 1 ficha en la misma casilla.";
                     else $mensaje="No hay ninguna ficha tuya en la casilla seleccionada.";
                 }else $mensaje = "No es tu turno.";
